refactor: split jpx CLI main.rs into focused modules
Extract 9 modules from the 3,359-line main.rs:
- args.rs: Args struct, ColorMode, Config, env/config loading
- output.rs: YAML, TOML, CSV, TSV, table, lines, JSON formatting
- discovery.rs: function listing, search, describe, similar
- stats.rs: data statistics, path listing
- bench.rs: benchmarking
- ops.rs: JSON diff/patch/merge, query validation
- explain.rs: AST printing, value description
- input.rs: file/stdin reading, JSON parsing, slurp mode
- streaming.rs: NDJSON streaming mode

main.rs is now 433 lines (orchestrator only). No public API or
behavior changes.

Closes #30
